+++ packages/types-augment-lookup/src/index.ts
@@ -1,0 +1,6 @@
+// Copyright 2017-2021 @polkadot/types-augment authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import './detectPackage';
+
+export * from './bundle';