+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Split `types-augment` into `registry` & `lookup` variants (better control)
+
+
 ## 7.0.2 Dec 21, 2021
 
 Upgrade priority: Low. Recommended for applications where collective derives with aliasses are being used.